Additional fixes
